refactor: move CompletionAdapter to new file
